requirements.txt
xgboost>=2.0.0
pymongo>=4.6.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
ta>=0.10.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
python-dotenv>=1.0.0
redis>=5.0.0
flask>=2.3.0